requirements.txt
streamlit==1.44.1
plotly==6.0.1
pandas==2.2.3
numpy==2.2.4
openai==1.75.0
python-dotenv==1.1.0
